+++ frontend/portal-funcionario/src/app/components/Card/CardFooter.js
