+++ frontend/portal-funcionario/src/app/components/Footer/Footer.js
