+++ frontend/portal-funcionario/src/app/components/Header/Header.js
