+++ frontend/portal-funcionario/src/app/feature/user/userDetailDialog.js
