+++ frontend/portal-funcionario/src/app/page.tsx
@@ -1,8 +1,28 @@
 import Image from "next/image";
+import Login from "./feature/login/login";
 
 export default function Home() {
+
   return (
     <div className="font-sans grid grid-rows-[20px_1fr_20px] items-center justify-items-center min-h-screen p-8 pb-20 gap-16 sm:p-20">
+      <div className="App">
+        <div>
+          { <div>"Bienvenido al portal del funcionario ":""</div>
+
+          }
+          <div>
+          { <Login />
+          }
+          </div>
+          <div>
+            {
+          
+            }
+          </div>
+          
+        </div>
+
+      </div>
       <main className="flex flex-col gap-[32px] row-start-2 items-center sm:items-start">
         <Image
           className="dark:invert"
